api/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
python-dotenv==1.0.0
pysnmplib==5.0.23
gunicorn==21.2.0